finalizado

+++ src/main/java/com/example/desafio_siad/repository/PessoaFisicaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.desafio_siad.domain.PessoaFisica;
 
-public interface PessoaFisicaRepository extends JpaRepository<PessoaFisica, Integer> {
+public interface PessoaFisicaRepository extends JpaRepository<PessoaFisica, Long> {
 
 }
