+++ src/main/java/com/example/desafio_siad/repository/PessoaJuridicaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.desafio_siad.domain.PessoaJuridica;
 
-public interface PessoaJuridicaRepository extends JpaRepository<PessoaJuridica, Integer> {
+public interface PessoaJuridicaRepository extends JpaRepository<PessoaJuridica, Long> {
 
 }
