+++ src/main/java/com/example/desafio_siad/repository/VendaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.desafio_siad.domain.Venda;
 
-public interface VendaRepository extends JpaRepository<Venda, Integer> {
+public interface VendaRepository extends JpaRepository<Venda, Long> {
 
 }
